Extract camera implementation
