+++ build_blockgame.bat
@@ -6,7 +6,7 @@ set "VSDEVCMD=C:\Program Files\Microsoft Visual Studio\2022\Community\Common7\To
 set "PROJECT_ROOT=%~dp0"
 set "INCLUDE_DIR=%PROJECT_ROOT%include"
 set "LIB_DIR=%PROJECT_ROOT%libs"
-set "SRC=src\main.cpp src\chunk_manager.cpp src\glad.c src\TextureLoader.cpp"
+set "SRC=src\main.cpp src\camera.cpp src\chunk_manager.cpp src\glad.c src\TextureLoader.cpp"
 set "OUT=blockgame.exe"
 set "GLFW_LIB=glfw3.lib"
 set "GLFW_DLL=glfw3.dll"
